+++ hydrodataset/__init__.py
@@ -28,8 +28,9 @@ from hydrodataset.configs.settings import (
 )
 from hydrodataset.configs.data_resolver import (
     READER_ALIASES,
-    resolve_data_path,
     DatasetResolutionError,
+    ResolverContext,
+    resolve_data_path,
 )
 
 # set some constants for datasets
@@ -40,7 +41,8 @@ LAMAH_REGIONS = ["CE"]
 MOPEX_REGIONS = ["CA"]
 REGIONS = CAMELS_REGIONS + LAMAH_REGIONS + MOPEX_REGIONS
 from .hydro_dataset import *
-#from .camelsh_kr import *
-#from .lamah_ce import *
-#from .camels_us import *
-#from .camels import *
+
+# from .camelsh_kr import *
+# from .lamah_ce import *
+# from .camels_us import *
+# from .camels import *